python-backend/app/utils/auth_middleware.py
# Authentication Middleware
from functools import wraps
from flask import request, jsonify
import jwt
import os


def token_required(f):
    """
    Decorator for protecting routes with JWT authentication
    Extracts user info from token and passes to route handler
    """
    @wraps(f)
    def decorated(*args, **kwargs):
        token = None
        
        # Get token from Authorization header
        if 'Authorization' in request.headers:
            auth_header = request.headers['Authorization']
            try:
                token = auth_header.split(' ')[1]  # Bearer <token>
            except IndexError:
                return jsonify({'error': 'Invalid token format'}), 401
        
        if not token:
            return jsonify({'error': 'Token is missing'}), 401
        
        try:
            # Decode token
            payload = jwt.decode(token, os.getenv('JWT_SECRET_KEY'), algorithms=['HS256'])
            current_user = {
                'id': payload['user_id'],
                'email': payload['email']
            }
        except jwt.ExpiredSignatureError:
            return jsonify({'error': 'Token has expired'}), 401
        except jwt.InvalidTokenError:
            return jsonify({'error': 'Invalid token'}), 401
        
        # Pass current user to route handler
        return f(current_user, *args, **kwargs)
    
    return decorated
