python-backend/requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-cors==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
werkzeug==3.0.1
pyjwt==2.8.0
redis==5.0.1
